Moved code to new file insertOnlineMemberData as counterpart to insertSomeHardcodedMemberData
