+++ Fotogroep Waalre/Fotogroep_Waalre_App.swift
@@ -25,7 +25,7 @@ struct FotogroepWaalreApp: App {
         UserDefaults.standard.set(Bundle.main.fullVersion, forKey: "version_preference")
     }
 
-    static var antiZombiePinningOfMemberPortfolios: Set<MemberPortfolio> = [] // hack to avoid zombies, global scope
+//    static var antiZombiePinningOfMemberPortfolios: Set<MemberPortfolio> = [] // hack to avoid zombies, needed? TODO
 
     var body: some Scene {
         WindowGroup {
